feat(phase-3.2): cite auto-discovery + sub-ms duration + fixtures
Three small fixes from the Phase 3.2 backlog plus one cleanup nit:

1. cite.md (item 2): replace MemWell-specific default reference set
   (events.jsonl, experiments.tsv) with generic SDD-aware auto-discovery
   that globs plan.md, data-model.md, research.md, quickstart.md,
   tasks.md, and contracts/**/*.md from the resolved feature dir.

2. duration_ms float precision (item 3): completion_gate runs in <1ms;
   int truncation rendered as "_duration: 0ms_". Switch to
   round(..., 1) for 0.1ms precision; result.to_json type hint becomes
   float; cli_output renders sub-ms as "0.3ms" and >=1ms as int.
   Adds 3 regression tests for the renderer.

3. Cleanup: drop unused tmp_path fixture from three citation_validator
   tests that don't write to disk (lines 154/186/200 per pyright).

Smoke: completion-gate now reports "duration_ms": 0.1 instead of 0.
Tests: 417 prior + 3 new = 420 passing.

+++ src/orca/core/result.py
@@ -14,7 +14,7 @@ class Ok(Generic[T]):
     value: T
     ok: Literal[True] = True
 
-    def to_json(self, *, capability: str, version: str, duration_ms: int) -> dict[str, Any]:
+    def to_json(self, *, capability: str, version: str, duration_ms: float) -> dict[str, Any]:
         return {
             "ok": True,
             "result": self.value,
@@ -31,7 +31,7 @@ class Err(Generic[E]):
     error: E
     ok: Literal[False] = False
 
-    def to_json(self, *, capability: str, version: str, duration_ms: int) -> dict[str, Any]:
+    def to_json(self, *, capability: str, version: str, duration_ms: float) -> dict[str, Any]:
         err_payload = (
             self.error.to_json()
             if hasattr(self.error, "to_json")
